[database] Added "lastEdited" and "inShoppingCart" fields to MKMMetaCard since they are already used in View Inventory
I left the names of the columns the same they were up to now in View Inventory for backwards compatibility, but some column names are inconsistent in whether they concatenate words or not, start with capital letter (inShoppingCart vs. Card Number for example)...would be nice to unify it in some way, though I don't care enough to do it.